chore: Split code files for util functions
